+++ app-local.env
@@ -1,4 +1,4 @@
-POSTGRES_DATABASE_URL=jdbc:postgresql://postgres:5432/study_kotlin
+POSTGRES_DATABASE_URL=jdbc:postgresql://postgres:5432/kotlin_kcl_sample
 POSTGRES_DATABASE_USER=test
 POSTGRES_DATABASE_PASSWORD=test
