--- backend/src/main/kotlin/com/study_kotlin/backend/BackendApplication.kt
@@ -1,13 +1,0 @@
-package com.study_kotlin.backend
-
-import org.springframework.boot.autoconfigure.SpringBootApplication
-import org.springframework.boot.runApplication
-import org.springframework.boot.context.properties.ConfigurationPropertiesScan
-
-@SpringBootApplication
-@ConfigurationPropertiesScan
-class BackendApplication
-
-fun main(args: Array<String>) {
-    runApplication<BackendApplication>(*args)
-}
