--- backend/src/main/kotlin/com/study_kotlin/backend/infrastructure/aws/CloudWatchConfig.kt
@@ -1,8 +1,0 @@
-package com.study_kotlin.backend.infrastructure.aws
-
-import org.springframework.boot.context.properties.ConfigurationProperties
-
-@ConfigurationProperties("cloudwatch")
-data class CloudWatchConfig (
-    val endpoint: String
-)
