--- backend/src/main/kotlin/com/study_kotlin/backend/infrastructure/aws/KinesisConfig.kt
@@ -1,10 +1,0 @@
-package com.study_kotlin.backend.infrastructure.aws
-
-import org.springframework.boot.context.properties.ConfigurationProperties
-
-@ConfigurationProperties("kinesis")
-data class KinesisConfig (
-    val endpoint: String,
-    val animalsStreamName: String,
-    val foodsStreamName: String
-)
