--- backend/src/main/kotlin/com/study_kotlin/backend/infrastructure/db/DbConfig.kt
@@ -1,10 +1,0 @@
-package com.study_kotlin.backend.infrastructure.db
-
-import org.springframework.boot.context.properties.ConfigurationProperties
-
-@ConfigurationProperties("postgres")
-data class DbConfig (
-    val url: String,
-    val username: String,
-    val password: String
-)
