--- backend/src/test/kotlin/com/study_kotlin/backend/BackendApplicationTests.kt
@@ -1,13 +1,0 @@
-package com.study_kotlin.backend
-
-import org.junit.jupiter.api.Test
-import org.springframework.boot.test.context.SpringBootTest
-
-@SpringBootTest
-class BackendApplicationTests {
-
-	@Test
-	fun contextLoads() {
-	}
-
-}
