+++ consumer-app/.gitattributes
