+++ consumer-app/src/main/kotlin/com/kotlin_kcl_sample/app/infrastructure/aws/DynamoDbConfig.kt
@@ -1,0 +1,8 @@
+package com.kotlin_kcl_sample.app.infrastructure.aws
+
+import org.springframework.boot.context.properties.ConfigurationProperties
+
+@ConfigurationProperties("dynamodb")
+data class DynamoDbConfig (
+    val endpoint: String
+)
