+++ consumer-app/src/main/resources/application.yaml
