+++ consumer-app/src/test/kotlin/com/kotlin_kcl_sample/app/SampleApplicationTests.kt
@@ -1,0 +1,13 @@
+package com.kotlin_kcl_sample.app
+
+import org.junit.jupiter.api.Test
+import org.springframework.boot.test.context.SpringBootTest
+
+@SpringBootTest
+class SampleApplicationTests {
+
+	@Test
+	fun contextLoads() {
+	}
+
+}
