+++ consumer-app/start.sh
