+++ infra/docker/postgres/init/init.sql
@@ -2,22 +2,22 @@ DO
 $$
 BEGIN
     IF
-EXISTS (SELECT FROM pg_database WHERE datname = 'study_kotlin_test') THEN
-        PERFORM pg_terminate_backend(pid) FROM pg_stat_activity WHERE datname = 'study_kotlin_test';
+EXISTS (SELECT FROM pg_database WHERE datname = 'kotlin_kcl_sample_test') THEN
+        PERFORM pg_terminate_backend(pid) FROM pg_stat_activity WHERE datname = 'kotlin_kcl_sample_test';
         DROP
-DATABASE study_kotlin_test;
+DATABASE kotlin_kcl_sample_test;
 END IF;
 END
 $$;
 
-CREATE DATABASE study_kotlin_test
+CREATE DATABASE kotlin_kcl_sample_test
     WITH OWNER = 'test'
     ENCODING = 'UTF8'
     LC_COLLATE = 'ja_JP.UTF-8'
     LC_CTYPE = 'ja_JP.UTF-8'
     TEMPLATE = template0;
 
-\c study_kotlin_test;
+\c kotlin_kcl_sample_test;
 
 CREATE TABLE IF NOT EXISTS kinesis_animals_data (
     id SERIAL PRIMARY KEY,
@@ -35,7 +35,7 @@ CREATE TABLE IF NOT EXISTS kinesis_foods_data (
     created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP
 );
 
-\c study_kotlin;
+\c kotlin_kcl_sample;
 
 CREATE TABLE IF NOT EXISTS kinesis_animals_data (
     id SERIAL PRIMARY KEY,
